include gstdio in order to silence some build warnings
git-svn-id: svn+ssh://svn.abisource.com/svnroot/enchant/trunk@22350 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ src/pwl.c
@@ -64,6 +64,7 @@
 #endif
 
 #include <glib.h>
+#include <glib/gstdio.h>
 
 #include "pwl.h"
 
@@ -1042,7 +1043,7 @@ static void enchant_trie_matcher_poppath(EnchantTrieMatcher* matcher,int num)
 static int edit_dist(const char* utf8word1, const char* utf8word2)
 {
 	gunichar * word1, * word2;
-	int len1, len2, cost, i, j;
+	glong len1, len2, cost, i, j;
 	int v1, v2, v3;
 	int* table;
